Fix -Wsign-compare in return

+++ domain/include/cstone/domain/domain.hpp
@@ -450,7 +450,8 @@ private:
     static bool sizesAllEqualTo(std::size_t value, Arrays&... arrays)
     {
         std::array<std::size_t, sizeof...(Arrays)> sizes{arrays.size()...};
-        return std::count(begin(sizes), end(sizes), value) == sizes.size();
+
+        return (size_t)std::count(begin(sizes), end(sizes), value) == sizes.size();
     }
 
     int myRank_;
